Move `CpuSet` out of `ostd::arch` and implement `this_cpu`
