+++ ostd/src/arch/x86/boot/mod.rs
@@ -22,6 +22,8 @@ mod linux_boot;
 mod multiboot;
 mod multiboot2;
 
+pub mod smp;
+
 use core::arch::global_asm;
 
 global_asm!(include_str!("boot.S"));
